+++ packages/seedless-onboarding-controller/src/SeedlessOnboardingController.test.ts
@@ -1,8 +1,5 @@
 import { keccak256AndHexify } from '@metamask/auth-network-utils';
-import {
-  deriveStateFromMetadata,
-  type Messenger,
-} from '@metamask/base-controller';
+import { deriveStateFromMetadata } from '@metamask/base-controller/next';
 import type { EncryptionKey } from '@metamask/browser-passworder';
 import {
   encrypt,
@@ -12,6 +9,7 @@ import {
   decryptWithKey as decryptWithKeyBrowserPassworder,
   importKey as importKeyBrowserPassworder,
 } from '@metamask/browser-passworder';
+import type { Messenger } from '@metamask/messenger';
 import {
   TOPRFError,
   type FetchAuthPubKeyResult,
@@ -54,9 +52,11 @@ import type {
   VaultEncryptor,
 } from './types';
 import type {
-  ExtractAvailableAction,
-  ExtractAvailableEvent,
-} from '../../base-controller/tests/helpers';
+  AllSeedlessOnboardingControllerActions,
+  AllSeedlessOnboardingControllerEvents,
+  baseMessengerName,
+  MockKeyringControllerMessenger,
+} from '../tests/__fixtures__/mockMessenger';
 import { mockSeedlessOnboardingMessenger } from '../tests/__fixtures__/mockMessenger';
 import {
   handleMockSecretDataGet,
@@ -123,9 +123,11 @@ type WithControllerCallback<ReturnValue, EKey> = ({
   initialState: SeedlessOnboardingControllerState;
   messenger: SeedlessOnboardingControllerMessenger;
   baseMessenger: Messenger<
-    ExtractAvailableAction<SeedlessOnboardingControllerMessenger>,
-    ExtractAvailableEvent<SeedlessOnboardingControllerMessenger>
+    typeof baseMessengerName,
+    AllSeedlessOnboardingControllerActions,
+    AllSeedlessOnboardingControllerEvents
   >;
+  keyringControllerMessenger: MockKeyringControllerMessenger;
   toprfClient: ToprfSecureBackup;
   mockRefreshJWTToken: jest.Mock;
   mockRevokeRefreshToken: jest.Mock;
@@ -185,7 +187,8 @@ async function withController<ReturnValue>(
 ) {
   const [{ ...rest }, fn] = args.length === 2 ? args : [{}, args[0]];
   const encryptor = new MockVaultEncryptor();
-  const { messenger, baseMessenger } = mockSeedlessOnboardingMessenger();
+  const { messenger, baseMessenger, keyringControllerMessenger } =
+    mockSeedlessOnboardingMessenger();
 
   const mockRefreshJWTToken = jest.fn().mockResolvedValue({
     idTokens: ['newIdToken'],
@@ -235,6 +238,7 @@ async function withController<ReturnValue>(
     initialState: controller.state,
     messenger,
     baseMessenger,
+    keyringControllerMessenger,
     toprfClient,
     mockRefreshJWTToken,
     mockRevokeRefreshToken,
@@ -5473,7 +5477,7 @@ describe('SeedlessOnboardingController', () => {
             deriveStateFromMetadata(
               controller.state,
               controller.metadata,
-              'anonymous',
+              'includeInDebugSnapshot',
             ),
           ).toMatchInlineSnapshot(`
             Object {